fix possible isspace() conversion warnings
git-svn-id: https://dev.csoft.net/agar/trunk@9914 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ gui/colors.c
@@ -77,7 +77,7 @@ AG_ColorFromString(const char *s, const AG_Color *pColor)
 
 	Strlcpy(buf, s, sizeof(buf));
 	
-	for (c = &buf[0]; *c != '\0' && isspace(*c); c++) {
+	for (c = &buf[0]; *c != '\0' && isspace((int)*c); c++) {
 		;;
 	}
 	switch (*c) {
